+++ derive/src/lib.rs
@@ -7,9 +7,9 @@ use proc_macro2::Span;
 use quote::quote;
 use syn::spanned::Spanned;
 use syn::{
-    parse_macro_input, parse_quote, Attribute, Data, DataEnum, DataStruct, DataUnion, DeriveInput,
-    Expr, ExprStruct, Field, Fields, FieldsNamed, Generics, Ident, ImplGenerics, IntSuffix, ItemFn,
-    Lit, Meta, MetaList, MetaNameValue, NestedMeta, Path, Stmt, Type, Visibility,
+    parse_macro_input, parse_quote, Attribute, BinOp, Data, DataEnum, DataStruct, DataUnion,
+    DeriveInput, Expr, ExprStruct, Field, Fields, FieldsNamed, Generics, Ident, ImplGenerics,
+    IntSuffix, ItemFn, Lit, Meta, MetaList, MetaNameValue, NestedMeta, Path, Type, Visibility,
 };
 
 use std::convert::TryFrom;
@@ -271,12 +271,11 @@ pub fn add(input: proc_macro::TokenStream) -> proc_macro::TokenStream {
 
 #[proc_macro_derive(AddAssign, attributes(modtype))]
 pub fn add_assign(input: proc_macro::TokenStream) -> proc_macro::TokenStream {
-    bin_assign_almost_transparent(
+    bin_assign(
         input,
         parse_quote!(AddAssign),
         parse_quote!(add_assign),
-        None,
-        |l, r, _| parse_quote!(#l += #r;),
+        parse_quote!(+),
     )
 }
 
@@ -292,12 +291,11 @@ pub fn sub(input: proc_macro::TokenStream) -> proc_macro::TokenStream {
 
 #[proc_macro_derive(SubAssign, attributes(modtype))]
 pub fn sub_assign(input: proc_macro::TokenStream) -> proc_macro::TokenStream {
-    bin_assign_almost_transparent(
+    bin_assign(
         input,
         parse_quote!(SubAssign),
         parse_quote!(sub_assign),
-        Some(parse_quote!(#[allow(clippy::suspicious_op_assign_impl)])),
-        |l, r, m| parse_quote!(#l -= #m + #r;),
+        parse_quote!(-),
     )
 }
 
@@ -313,12 +311,11 @@ pub fn mul(input: proc_macro::TokenStream) -> proc_macro::TokenStream {
 
 #[proc_macro_derive(MulAssign, attributes(modtype))]
 pub fn mul_assign(input: proc_macro::TokenStream) -> proc_macro::TokenStream {
-    bin_assign_almost_transparent(
+    bin_assign(
         input,
         parse_quote!(MulAssign),
         parse_quote!(mul_assign),
-        None,
-        |l, r, _| parse_quote!(#l *= #r;),
+        parse_quote!(*),
     )
 }
 
@@ -377,17 +374,12 @@ pub fn div(input: proc_macro::TokenStream) -> proc_macro::TokenStream {
 
 #[proc_macro_derive(DivAssign, attributes(modtype))]
 pub fn div_assign(input: proc_macro::TokenStream) -> proc_macro::TokenStream {
-    bin_assign(input, parse_quote!(DivAssign), |input, rhs_ty| {
-        let Input { std, .. } = input;
-        parse_quote! {
-            #[inline]
-            fn div_assign(&mut self, rhs: #rhs_ty) {
-                fn static_assert_copy<T: #std::marker::Copy>() {}
-                static_assert_copy::<Self>();
-                *self = *self / rhs;
-            }
-        }
-    })
+    bin_assign(
+        input,
+        parse_quote!(DivAssign),
+        parse_quote!(div_assign),
+        parse_quote!(/),
+    )
 }
 
 #[proc_macro_derive(Rem, attributes(modtype))]
@@ -420,80 +412,11 @@ pub fn rem(input: proc_macro::TokenStream) -> proc_macro::TokenStream {
 
 #[proc_macro_derive(RemAssign, attributes(modtype))]
 pub fn rem_assign(input: proc_macro::TokenStream) -> proc_macro::TokenStream {
-    bin_assign(input, parse_quote!(RemAssign), |input, rhs_ty| {
-        let Input { std, .. } = input;
-
-        parse_quote! {
-            #[inline]
-            fn rem_assign(&mut self, rhs: #rhs_ty) {
-                fn static_assert_copy<T: #std::marker::Copy>() {}
-                static_assert_copy::<Self>();
-                *self = *self % rhs;
-            }
-        }
-    })
-}
-
-#[proc_macro_derive(BitAnd, attributes(modtype))]
-pub fn bit_and(input: proc_macro::TokenStream) -> proc_macro::TokenStream {
-    bin_almost_transparent(
+    bin_assign(
         input,
-        parse_quote!(BitAnd),
-        parse_quote!(bitand),
-        |l, r, _| parse_quote!(#l & #r),
-    )
-}
-
-#[proc_macro_derive(BitAndAssign, attributes(modtype))]
-pub fn bit_and_assgin(input: proc_macro::TokenStream) -> proc_macro::TokenStream {
-    bin_assign_almost_transparent(
-        input,
-        parse_quote!(BitAndAssign),
-        parse_quote!(bitand_assign),
-        None,
-        |l, r, _| parse_quote!(#l &= #r;),
-    )
-}
-
-#[proc_macro_derive(BitOr, attributes(modtype))]
-pub fn bit_or(input: proc_macro::TokenStream) -> proc_macro::TokenStream {
-    bin_almost_transparent(
-        input,
-        parse_quote!(BitOr),
-        parse_quote!(bitor),
-        |l, r, _| parse_quote!(#l | #r),
-    )
-}
-
-#[proc_macro_derive(BitOrAssign, attributes(modtype))]
-pub fn bit_or_assign(input: proc_macro::TokenStream) -> proc_macro::TokenStream {
-    bin_assign_almost_transparent(
-        input,
-        parse_quote!(BitOrAssign),
-        parse_quote!(bitor_assign),
-        None,
-        |l, r, _| parse_quote!(#l |= #r;),
-    )
-}
-
-#[proc_macro_derive(BitXor, attributes(modtype))]
-pub fn bit_xor(input: proc_macro::TokenStream) -> proc_macro::TokenStream {
-    bin_almost_transparent(
-        input,
-        parse_quote!(BitXor),
-        parse_quote!(bitxor),
-        |l, r, _| parse_quote!(#l ^ #r),
-    )
-}
-
-#[proc_macro_derive(BitXorAssign, attributes(modtype))]
-pub fn bit_xor_assign(input: proc_macro::TokenStream) -> proc_macro::TokenStream {
-    bin_assign_almost_transparent(
-        input,
-        parse_quote!(BitXorAssign),
-        parse_quote!(bitxor_assign),
-        None,
-        |l, r, _| parse_quote!(#l ^= #r;),
+        parse_quote!(RemAssign),
+        parse_quote!(rem_assign),
+        parse_quote!(%),
     )
 }
 
@@ -574,6 +497,88 @@ pub fn bounded(input: proc_macro::TokenStream) -> proc_macro::TokenStream {
             fn max_value() -> Self {
                 let minus_1 = #modulus - <#field_ty as #num_traits::One>::one();
                 <Self as #std::convert::From<#field_ty>>::from(minus_1)
+            }
+        }
+    )
+    .into()
+}
+
+#[proc_macro_derive(CheckedAdd, attributes(modtype))]
+pub fn checked_add(input: proc_macro::TokenStream) -> proc_macro::TokenStream {
+    checked_bin(
+        input,
+        parse_quote!(CheckedAdd),
+        parse_quote!(checked_add),
+        false,
+        parse_quote!(+),
+    )
+}
+
+#[proc_macro_derive(CheckedSub, attributes(modtype))]
+pub fn checked_sub(input: proc_macro::TokenStream) -> proc_macro::TokenStream {
+    checked_bin(
+        input,
+        parse_quote!(CheckedSub),
+        parse_quote!(checked_sub),
+        false,
+        parse_quote!(-),
+    )
+}
+
+#[proc_macro_derive(CheckedMul, attributes(modtype))]
+pub fn checked_mul(input: proc_macro::TokenStream) -> proc_macro::TokenStream {
+    checked_bin(
+        input,
+        parse_quote!(CheckedMul),
+        parse_quote!(checked_mul),
+        false,
+        parse_quote!(*),
+    )
+}
+
+#[proc_macro_derive(CheckedDiv, attributes(modtype))]
+pub fn checked_div(input: proc_macro::TokenStream) -> proc_macro::TokenStream {
+    checked_bin(
+        input,
+        parse_quote!(CheckedDiv),
+        parse_quote!(checked_div),
+        true,
+        parse_quote!(/),
+    )
+}
+
+#[proc_macro_derive(CheckedRem, attributes(modtype))]
+pub fn checked_rem(input: proc_macro::TokenStream) -> proc_macro::TokenStream {
+    checked_bin(
+        input,
+        parse_quote!(CheckedRem),
+        parse_quote!(checked_rem),
+        true,
+        parse_quote!(%),
+    )
+}
+
+#[proc_macro_derive(CheckedNeg, attributes(modtype))]
+pub fn checked_neg(input: proc_macro::TokenStream) -> proc_macro::TokenStream {
+    let input = parse_macro_input!(input as DeriveInput);
+    let Input {
+        std,
+        num_traits,
+        struct_ident,
+        generics,
+        ..
+    } = try_syn!(Input::try_from(input));
+    let (impl_generics, ty_generics, where_clause) = generics.split_for_impl();
+
+    quote!(
+        impl#impl_generics #num_traits::CheckedNeg for #struct_ident#ty_generics
+            #where_clause
+        {
+            #[inline]
+            fn checked_neg(&self) -> #std::option::Option<Self> {
+                fn static_assert_copy<T: #std::marker::Copy>() {}
+                static_assert_copy::<Self>();
+                Some(-*self)
             }
         }
     )
@@ -1036,42 +1041,11 @@ fn bin(
     ret.into()
 }
 
-fn bin_assign_almost_transparent(
-    input: proc_macro::TokenStream,
-    trait_ident: Ident,
-    fn_ident: Ident,
-    attr: Option<Attribute>,
-    op: fn(&Expr, &Expr, &Expr) -> Stmt,
-) -> proc_macro::TokenStream {
-    bin_assign(input, trait_ident, |input, rhs_ty| {
-        let Input {
-            modulus,
-            field_ident,
-            ..
-        } = input;
-
-        let stmt = op(
-            &parse_quote!(self.#field_ident),
-            &parse_quote!(rhs.#field_ident),
-            &modulus,
-        );
-        parse_quote! {
-            #attr
-            #[inline]
-            fn #fn_ident(&mut self, rhs: #rhs_ty) {
-                #stmt
-                if self.#field_ident >= #modulus {
-                    self.#field_ident %= #modulus;
-                }
-            }
-        }
-    })
-}
-
 fn bin_assign(
     input: proc_macro::TokenStream,
     trait_ident: Ident,
-    derive_fn: impl Fn(&Input, &Type) -> ItemFn,
+    fn_ident: Ident,
+    bin_op: BinOp,
 ) -> proc_macro::TokenStream {
     let input = parse_macro_input!(input as DeriveInput);
     let input = try_syn!(Input::try_from(input));
@@ -1084,27 +1058,26 @@ fn bin_assign(
     } = &input;
     let (impl_generics, ty_generics, where_clause) = generics.split_for_impl();
 
-    let derive = |impl_generics: &ImplGenerics, rhs_ty: Type| -> _ {
-        let item_fn = derive_fn(&input, &rhs_ty);
+    let derive = |rhs_ty: Type, rhs_deref: bool| -> _ {
+        let star_token = if rhs_deref { quote!(*) } else { quote!() };
         quote! {
             impl#impl_generics #std::ops::#trait_ident<#rhs_ty> for #struct_ident#ty_generics
             #where_clause
             {
-                #item_fn
+                #[inline]
+                fn #fn_ident(&mut self, rhs: #rhs_ty) {
+                    fn static_assert_copy<T: #std::marker::Copy>() {}
+                    static_assert_copy::<Self>();
+                    *self = *self #bin_op #star_token rhs;
+                }
             }
         }
     };
 
-    let mut ret = derive(&impl_generics, parse_quote!(#struct_ident#ty_generics));
-
+    let mut ret = derive(parse_quote!(Self), false);
     if *moving_ops_for_ref {
-        let (impl_generics, _, _) = generics.split_for_impl();
-        ret.extend(derive(
-            &impl_generics,
-            parse_quote!(&'_ #struct_ident#ty_generics),
-        ));
+        ret.extend(derive(parse_quote!(&'_ Self), true));
     }
-
     ret.into()
 }
 
@@ -1143,6 +1116,50 @@ fn identity(
             #[inline]
             fn #is(&self) -> bool {
                 <#field_ty as #num_traits::#trait_ident>::#is(&self.#field_ident)
+            }
+        }
+    )
+    .into()
+}
+
+fn checked_bin(
+    input: proc_macro::TokenStream,
+    trait_ident: Ident,
+    fn_ident: Ident,
+    return_none_if_rhs_is_zero: bool,
+    op: BinOp,
+) -> proc_macro::TokenStream {
+    let input = parse_macro_input!(input as DeriveInput);
+    let Input {
+        std,
+        num_traits,
+        struct_ident,
+        generics,
+        ..
+    } = try_syn!(Input::try_from(input));
+    let (impl_generics, ty_generics, where_clause) = generics.split_for_impl();
+
+    let expr: Expr = if return_none_if_rhs_is_zero {
+        parse_quote! {
+            if <Self as #num_traits::Zero>::is_zero(rhs) {
+                None
+            } else {
+                Some(*self #op *rhs)
+            }
+        }
+    } else {
+        parse_quote!(Some(*self #op *rhs))
+    };
+
+    quote!(
+        impl#impl_generics #num_traits::#trait_ident for #struct_ident#ty_generics
+            #where_clause
+        {
+            #[inline]
+            fn #fn_ident(&self, rhs: &Self) -> #std::option::Option<Self> {
+                fn static_assert_copy<T: #std::marker::Copy>() {}
+                static_assert_copy::<Self>();
+                #expr
             }
         }
     )